CanSee mhetod-enemy

+++ enemy.h
@@ -6,8 +6,8 @@ class Enemy : public Creature
 {
 protected:
     enum move_dir{up,right,down,left};
-    int dst_X;
-    int dst_Y;
+    float dst_X;
+    float dst_Y;
     bool CanSee(Board *_map);
 public:
     Enemy(Board *_map, float _speed, int _maxHp, int _width, int _height);
